separate into download utils file
